Prepare reproducible manuscript experiments and fix backend audit correctness
Add verified selection, blocked sparse witness graphs, native FAISS batching,
and a reproducible CPU experiment matrix. Fix input mutation, capped-search
metadata, candidate verification, color audit precision, and exactness gates.

+++ src/fast_ballmapper/audit.py
@@ -275,7 +275,10 @@ def _audit_mean_colors(
     approximate_sets: list[set[int]],
     values: np.ndarray,
 ) -> ColorAudit:
-    scalar_values = np.asarray(values)
+    # Accumulate means in the same float64 precision used for oscillations and
+    # bounds. Casting a float32 mean afterwards can create spurious violations
+    # when the finite-set color bound is attained exactly.
+    scalar_values = np.asarray(values, dtype=np.float64)
     if scalar_values.ndim != 1:
         raise ValueError("values must be a one-dimensional array of scalar colors.")
